Update: reuse _write_function_code for both functions and methods
Previously code was duplicated in 2 places - functions and methods. Which resulted in `/` not being processed correctly for method (e.g. `test(a, b, /, c)` was written as `test(a, b, c)`.

No functional changes here mostly, the only change is that methods now also have `current_status` handling. There's still need to sort out differences in `# documentation`, but leaving it as is for now to fix currently breaking stub generation.